--- lldb/test/API/functionalities/data-formatter/swift/url/Makefile
@@ -1,3 +1,0 @@
-SWIFT_SOURCES := main.swift
-
-include Makefile.rules
