+++ lldb/test/API/lang/swift/foundation_value_types/url/TestSwiftFoundationTypeURL.py
@@ -9,9 +9,54 @@
 # See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
 #
 # ------------------------------------------------------------------------------
-import lldbsuite.test.lldbinline as lldbinline
-from lldbsuite.test.decorators import *
+"""
+Test Foundation.URL summary strings.
+"""
 
-lldbinline.MakeInlineTest(__file__, globals(),
-                          decorators=[swiftTest,skipUnlessDarwin,
-                            expectedFailureAll(archs=['arm64_32'], bugnumber="<rdar://problem/58065423>")])
+import lldb
+import sys
+
+from lldbsuite.test.lldbtest import *
+from lldbsuite.test.decorators import *
+import lldbsuite.test.lldbutil as lldbutil
+
+
+class TestCase(TestBase):
+    @skipUnlessFoundation
+    @swiftTest
+    def test_swift_url_formatters(self):
+        """Test URL summary strings."""
+        self.build()
+
+        foundation = "Foundation" if sys.platform == "darwin" else "FoundationEssentials"
+
+        lldbutil.run_to_source_breakpoint(
+            self, "break here", lldb.SBFileSpec("main.swift")
+        )
+
+        self.expect(
+            "frame var url",
+            startstr=f'({foundation}.URL?) url = "https://www.example.com/path?query#fragment"',
+        )
+        self.expect(
+            "expression -d run -- url",
+            startstr=f'({foundation}.URL?) $R0 = "https://www.example.com/path?query#fragment"',
+        )
+
+        self.expect(
+            "frame var relativeURL",
+            startstr=f'({foundation}.URL?) relativeURL = "relative -- https://www.example.com/"',
+        )
+        self.expect(
+            "expression -d run -- relativeURL",
+            startstr=f'({foundation}.URL?) $R1 = "relative -- https://www.example.com/"',
+        )
+
+        self.expect(
+            "frame var g_url",
+            startstr=f'({foundation}.URL) g_url = "http://www.apple.com"',
+        )
+        self.expect(
+            "expression -d run -- g_url",
+            startstr=f'({foundation}.URL) $R2 = "http://www.apple.com"',
+        )
